add support for nested resources (sub-resources)
SousVide is now aware of sub-resources and will process these the same
way it does with top-level resources. Previously sub-resources were
simply ignored.

This requires additional handler state:

* `@nested_queue`

This is an array that acts as a stack. If it's not empty the last
element is a parent of currently processed resource. Resources are added
to it based on order of events received.

* `@current_event` and `@previous_event`

These represent events emitted by Chef, regardless if or how SousVide
handles them. Out of bound events that are generally ignored will still
populate these variables. Order of events is significant because it
allows SousVide to detect when Chef enters nested resources.

* `@current_action` represents currently processed action. It is
significant because `resource_completed` event does not pass an action.

Additionally `TrackedResource` got `nest_level` field that represents
nesting depth. `nest_level` combined with `execution_order` is
sufficient to determine a parent resource.

When SousVide receives subsequent `:resource_action_start` before
`:resource_completed` it pushes current resource to the `@nested_queue`
and increments nest level. When it receives out-of-bound event later (no
`@processing_now` it will pop the top resource from `@nested_queue`.
Past that the processing itself remains unchanged.